qcacld-3.0: Handle UBSAN OOB access issue in vendor scan
The kernel cfg80211_scan_request data structure has flexible
length array for channels. Currently in the driver, based on
number of channels and ssid information, a single large
continuous block of memory is allocated.

The memory alignment is channels[], ssids[] and ie[]. But from
kernel 6.16, this direct access of request->channels[] is treated
as OOB access by UBSAN.

To avoid this issue and ensure proper memory alignment, use pointer
arithmetic and offsetof() to determine the arrays in request struct.

Change-Id: I6c81c894f4eb9bf7338013e6f859b863fb1ebe89
CRs-Fixed: 4340263